requirements.txt
certifi==2022.6.15
charset-normalizer==2.1.0
commonmark==0.9.1
idna==3.3
Pygments==2.12.0
requests==2.28.1
rich==12.4.4
urllib3==1.26.9
wcmatch==8.4
GitPython==3.1.27
